add file_list_page, refactor mailman_client
